renamed Wrapper class to base.py
